\encoding{UTF-8}
\name{coralheads}
\Rdversion{1.1}
\alias{coralheads}
\docType{data}
\title{
Richness and Abundance of Decapod Crustaceans on Coral Heads
}
\description{
The data set contains observations from Abele and Patton (1976).
\code{A} is the size of the coral heads in \eqn{cm^3},
\code{S} is the number of decapod crustacean species,
\code{N} is the number of decapod crustacean individuals on the coral heads.
}
\usage{data(coralheads)}
\source{
Abele, L. G. and W. K. Patton. 1976.
The size of coral heads and the community biology of associated decapod crustaceans.
\emph{Journal of Biogeography} \bold{3}, 35--47.
}
\examples{
\dontrun{
data(coralheads)
summary(coralheads)
str(coralheads)
}
}
\keyword{datasets}
