\encoding{UTF-8}
\name{diagnose}
\Rdversion{1.1}
\alias{diagnose}
\title{
Diagnose R Objects
}
\description{
Generic function to diagnose R objects, such as MCMC convergence, etc.
}
\usage{
diagnose(x, ...)
}
\arguments{
  \item{x}{
An R object.
}
  \item{\dots}{
Further arguments.
}
}
\author{
\enc{P\'eter S\'olymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\keyword{ methods }

