\encoding{UTF-8}
\name{regmod}
\Rdversion{1.1}
\alias{regmod}
\docType{data}
\title{
Exemplary MCMC list object
}
\description{
This data set was done via the code for \code{\link{jags.fit}} function.
}
\usage{data(regmod)}
\source{
See example of \code{\link{jags.fit}}.
}
\examples{
\dontrun{data(regmod)
summary(regmod)
plot(regmod)
}
}
\keyword{datasets}
