\name{report.mcmc.list}
\Rdversion{1.1}
\alias{report.mcmc.list}
\title{
Conveniently calculating statistics for \code{mcmc.list} objects.
}
\usage{
report.mcmc.list(x, FUN, ...)
}
\arguments{
  \item{x}{
An object of class \code{mcmc.list}.
}
  \item{FUN}{
A function to be used in the calculations.
}
  \item{\dots}{
Other arguments passed to \code{FUN}.
}
}
\value{
Statistic value based on \code{FUN}.
}
\author{
\enc{Péter Sólymos}{Peter Solymos}, \email{solymos@ualberta.ca}
}
\keyword{ methods }
\keyword{ utilities }
